% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stats.R
\name{f.stats}
\alias{f.stats}
\title{Frequency Vector Statistics}
\usage{
f.stats(f)
}
\arguments{
\item{f}{a vector of species frequencies where \code{f[i]} is the number 
of species represented by only \code{i} samples.}
}
\value{
a vector of the number of observed species (\code{s.obs}), 
  and the total number of samples (\code{n}).
}
\description{
Number of observed species and samples in species frequency 
  vector.
}
\examples{
data(osa.second.growth)
f <- expand.freqs(osa.second.growth)
f.stats(f)

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
