% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sprex-package.R
\docType{data}
\name{osa.second.growth}
\alias{osa.second.growth}
\title{Osa second growth}
\usage{
data(osa.second.growth)
}
\description{
Osa second growth
}
\references{
Janzen DH (1973) Sweep samples of tropical foliage insects:
  description of study sites, with data on species abundances and size
  distributions. Ecology 54:659-86.
}
\keyword{datasets}

