% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.to.sample.freq.R
\name{species.to.sample.freq}
\alias{species.to.sample.freq}
\title{Create Vector of Sample Frequencies}
\usage{
species.to.sample.freq(f)
}
\arguments{
\item{f}{a vector of species frequencies where \code{f[i]} is the 
number of species represented by only \code{i} samples.}
}
\value{
a vector(\code{x}) where \code{x[i]} is of the number of samples 
  in the \code{i}-th species.
}
\description{
Create vector of sample frequencies from vector of 
  species frequencies.
}
\examples{
data(osa.old.growth)
f <- expand.freqs(osa.old.growth)
x <- species.to.sample.freq(f)
print(f)
print(x)

}
\seealso{
sample.to.species.freq
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
