% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sprex-package.R
\docType{data}
\name{osa.old.growth}
\alias{osa.old.growth}
\title{Osa old growth}
\usage{
data(osa.old.growth)
}
\description{
Osa old growth
}
\references{
Janzen DH (1973) Sweep samples of tropical foliage insects:
  effects of seasons, vegetation types, elevation, time of day,
  and insularity. Ecology 54:687-708.
}
\keyword{datasets}

